refactor: cleanup surrounding events
