+++ lib/store/events/receipts/storage.dart
@@ -6,7 +6,7 @@ import 'package:syphon/global/algos.dart';
 import 'package:syphon/global/print.dart';
 import 'package:syphon/storage/constants.dart';
 import 'package:syphon/store/events/ephemeral/m.read/model.dart';
-import 'package:syphon/store/events/redaction/model.dart';
+import 'package:syphon/store/events/redactions/model.dart';
 
 ///
 /// Save Receipts
