--- lib/store/events/redaction/selectors.dart
@@ -1,3 +1,0 @@
-import 'package:syphon/store/events/messages/model.dart';
-import 'package:syphon/store/events/reactions/model.dart';
-import 'package:syphon/store/index.dart';
