+++ lib/store/events/redactions/model.dart
